feat : #28/Exception Global 클래스 분리
