--- src/main/java/com/sparta/delivery_app/common/JpaAuditingConfig.java
@@ -1,9 +1,0 @@
-package com.sparta.delivery_app.common;
-
-import org.springframework.context.annotation.Configuration;
-import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
-
-@Configuration
-@EnableJpaAuditing
-public class JpaAuditingConfig {
-}
