+++ src/main/java/com/sparta/delivery_app/common/exceptionhandler/GlobalExceptionHandler.java
@@ -1,9 +1,39 @@
 package com.sparta.delivery_app.common.exceptionhandler;
 
+import com.sparta.delivery_app.common.exception.errorcode.ErrorCode;
+import com.sparta.delivery_app.common.globalResponse.ErrorResponse;
+import com.sparta.delivery_app.common.globalcustomexception.global.GlobalDuplicatedException;
+import com.sparta.delivery_app.common.globalcustomexception.global.GlobalServerException;
 import lombok.extern.slf4j.Slf4j;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.RestControllerAdvice;
+
+import javax.naming.AuthenticationException;
 
 @Slf4j
 @RestControllerAdvice
 public class GlobalExceptionHandler {
+
+    /**
+     * 서버 500에러 발생
+     */
+    @ExceptionHandler(GlobalServerException.class)
+    protected ResponseEntity<ErrorResponse> globalServerException(GlobalServerException e) {
+        ErrorCode errorCode = e.getErrorCode();
+
+        return ResponseEntity.status(errorCode.getHttpStatusCode())
+                .body(ErrorResponse.of(errorCode));
+    }
+
+    @ExceptionHandler(AuthenticationException.class)
+    protected void AuthenticationException(AuthenticationException e) {
+        String message = e.getMessage();
+        log.info(message);
+        log.info("asdfasfasfasdfasfasfasf");
+
+//        return ResponseEntity.status(errorCode.getHttpStatusCode())
+//                .body(ErrorResponse.of(errorCode));
+    }
+
 }
