+++ src/main/java/com/sparta/delivery_app/common/globalcustomexception/global/GlobalDuplicatedException.java
@@ -1,13 +1,14 @@
-package com.sparta.delivery_app.common.globalcustomexception;
+package com.sparta.delivery_app.common.globalcustomexception.global;
 
 import com.sparta.delivery_app.common.exception.errorcode.ErrorCode;
 import lombok.Getter;
 
 @Getter
-public class GlobalNotFoundException extends RuntimeException {
+public class GlobalDuplicatedException extends RuntimeException {
+
     private final ErrorCode errorCode;
 
-    public GlobalNotFoundException(ErrorCode errorCode) {
+    public GlobalDuplicatedException(ErrorCode errorCode) {
         super(errorCode.getDescription());
         this.errorCode = errorCode;
     }
