+++ src/main/java/com/sparta/delivery_app/common/globalcustomexception/global/GlobalMismatchException.java
@@ -1,0 +1,14 @@
+package com.sparta.delivery_app.common.globalcustomexception.global;
+
+import com.sparta.delivery_app.common.exception.errorcode.ErrorCode;
+import lombok.Getter;
+
+@Getter
+public class GlobalMismatchException extends RuntimeException {
+    private final ErrorCode errorCode;
+
+    public GlobalMismatchException(ErrorCode errorCode) {
+        super(errorCode.getDescription());
+        this.errorCode = errorCode;
+    }
+}
