+++ src/main/java/com/sparta/delivery_app/common/globalcustomexception/global/GlobalServerException.java
@@ -1,14 +1,13 @@
-package com.sparta.delivery_app.common.globalcustomexception;
+package com.sparta.delivery_app.common.globalcustomexception.global;
 
 import com.sparta.delivery_app.common.exception.errorcode.ErrorCode;
 import lombok.Getter;
 
 @Getter
-public class GlobalDuplicatedException extends RuntimeException {
-
+public class GlobalServerException extends RuntimeException {
     private final ErrorCode errorCode;
 
-    public GlobalDuplicatedException(ErrorCode errorCode) {
+    public GlobalServerException(ErrorCode errorCode) {
         super(errorCode.getDescription());
         this.errorCode = errorCode;
     }
