+++ src/main/resources/application.yml
@@ -27,3 +27,8 @@ spring:
 
 logging.level:
   root: info
+
+---
+
+jwt:
+  secret-key: ${JWT_SECRET_KEY}
